pyproject.toml
[tool.poetry]
name = "blacklist"
version = "0.1.0"
description = ""
authors = ["sonnya <sofi.kravchenko@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.116.1"
pydantic = "^2.11.7"
pytest = "^8.4.1"
tortoise-orm = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
